+++ pkg/services/ngalert/provisioning/errors.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 
 	"github.com/grafana/grafana/pkg/apimachinery/errutil"
-	"github.com/grafana/grafana/pkg/services/ngalert/models"
 )
 
 var ErrValidation = fmt.Errorf("invalid object specification")
@@ -15,11 +14,6 @@ var ErrPermissionDenied = errors.New("permission denied")
 var (
 	ErrNoAlertmanagerConfiguration  = errutil.Internal("alerting.notification.configMissing", errutil.WithPublicMessage("No alertmanager configuration present in this organization"))
 	ErrBadAlertmanagerConfiguration = errutil.Internal("alerting.notification.configCorrupted").MustTemplate("Failed to unmarshal the Alertmanager configuration", errutil.WithPublic("Current Alertmanager configuration in the storage is corrupted. Reset the configuration or rollback to a recent valid one."))
-
-	ErrProvenanceChangeNotAllowed = errutil.Forbidden("alerting.notifications.invalidProvenance").MustTemplate(
-		"Resource with provenance status '{{ .Public.SourceProvenance }}' cannot be managed via API that handles resources with provenance status '{{ .Public.TargetProvenance }}'",
-		errutil.WithPublic("Resource with provenance status '{{ .Public.SourceProvenance }}' cannot be managed via API that handles resources with provenance status '{{ .Public.TargetProvenance }}'. You must use appropriate API to manage this resource"),
-	)
 
 	ErrVersionConflict = errutil.Conflict("alerting.notifications.conflict")
 
@@ -52,20 +46,4 @@ func MakeErrTimeIntervalInvalid(err error) error {
 	}
 
 	return ErrTimeIntervalInvalid.Build(data)
-}
-
-func MakeErrProvenanceChangeNotAllowed(from, to models.Provenance) error {
-	if to == "" {
-		to = "none"
-	}
-	if from == "" {
-		from = "none"
-	}
-	data := errutil.TemplateData{
-		Public: map[string]interface{}{
-			"TargetProvenance": to,
-			"SourceProvenance": from,
-		},
-	}
-	return ErrProvenanceChangeNotAllowed.Build(data)
 }
