+++ pkg/services/ngalert/provisioning/validation/provenance.go
@@ -1,12 +1,12 @@
-package provisioning
+package validation
 
 import (
 	"github.com/grafana/grafana/pkg/services/ngalert/models"
 )
 
-// canUpdateProvenanceInRuleGroup checks if a provenance can be updated for a rule group and its alerts.
+// CanUpdateProvenanceInRuleGroup checks if a provenance can be updated for a rule group and its alerts.
 // ReplaceRuleGroup function intends to replace an entire rule group: inserting, updating, and removing rules.
-func canUpdateProvenanceInRuleGroup(storedProvenance, provenance models.Provenance) bool {
+func CanUpdateProvenanceInRuleGroup(storedProvenance, provenance models.Provenance) bool {
 	return storedProvenance == provenance ||
 		storedProvenance == models.ProvenanceNone ||
 		(storedProvenance == models.ProvenanceAPI && provenance == models.ProvenanceNone)
