+++ pkg/services/ngalert/provisioning/validation/provenance_test.go
@@ -1,4 +1,4 @@
-package provisioning
+package validation
 
 import (
 	"fmt"
